feat(engine): LoopNode — run body until/max_iterations with break (T034) (#34)
* feat(engine): LoopNode — run body until/max_iterations with break (T034)

Add LoopNode: runs the loop body via the scheduler each iteration until
the `until` post-condition is true or `max_iterations` is reached; a body
`break` exits early; `outputs` CEL expressions are evaluated at exit and
exposed via `needs.<loop_id>.output`. Exposes `scope` (iteration, nodes,
needs, outer) to body nodes, `until`, and `outputs`.

- schema: `loop.max_iterations` constrained to an integer
- scheduler: expose accumulated completed-node results as
  `SchedulerResult.nodeResults`
- workflow: register LoopNode
- `scope.nodes` is a per-iteration snapshot (absent nodes are undefined;
  guard with CEL `has()`) — no cross-iteration overlay


* test(engine): use real BreakNode in LoopNode break tests

Replace the dual-role FactoryNode stopper (completed one iteration, break the next) with a real BreakNode guarded by if: scope.iteration >= 1 in the two scope.nodes break tests; optional now uses a real if instead of overriding evaluateIf. Behavior and assertions unchanged.


* test(engine): consolidate LoopNode break tests into one partial-snapshot test

Collapse two overlapping 'on break' tests into one that exercises a genuine partial run: an 'after' node downstream of the break runs while the break is skipped (a skipped break does not cascade its skip) and is cancelled when it fires, so it is absent from the partial snapshot rather than backfilled. Asserts completed/skipped/cut-off node presence, total_iterations, and the after run count.


* fix(engine): propagate cancellation into a loop's inner scheduler (PR review)

Thread an optional external AbortSignal through SchedulerOptions and compose it with the scheduler's internal controller (AbortSignal.any), so a LoopNode cancelled mid-iteration aborts its in-flight body nodes and returns ENGINE_NODE_CANCELLED instead of a body failure derived from the aborted run. Backward-compatible: the top-level workflow passes no signal. Adds a deterministic regression test.

Also: max_iterations doc says until must evaluate to 'true' (not 'truthy'), and corrects the lastIterationNodes comment (it also feeds outputs after the loop, not only the next iteration's scope).


---------

+++ src/core/engine/schema.ts
@@ -138,7 +138,7 @@ export const LoopNodeSchema = BaseNodeSchema.extend({
   loop: z
     .object({
       until: z.string().optional(),
-      max_iterations: z.number().min(1).optional(),
+      max_iterations: z.number().int().min(1).optional(),
       nodes: z.array(NodeSchema).min(1),
       outputs: z.record(z.string(), z.string()).optional(),
     })
